requirements.txt
selenium>=4.15.0
anthropic>=0.40.0
PyPDF2>=3.0.0
python-docx>=1.1.0
